requirements.txt
streamlit
openai
faiss-cpu
scikit-learn
pandas
numpy
tiktoken
